+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support direct-access mode on AQT devices (#164)
 * Support up to 2000 shots per circuit on cloud resources (#165)
+* Fix result timeout propagation in direct-access mode (#171)
 
 ## qiskit-aqt-provider v1.5.0